[refactor] file hierachy restructuring
