+++ website/src/pages/accounts/create-account.tsx
@@ -4,8 +4,8 @@ import AccountsPageContainer from '../../components/accounts-page-container';
 import { Notification,  NOTIFICATION_STYLE_ERROR } from '../../components/notification';
 import { CreateAccountPOST } from '../../models/http-requests';
 import { CREATE_ACCOUNT_END_POINT, http_post, HTTP_SUCCESS, GetHttpRequestDisplayName } from '../../services/http-service';
-import { FormFieldParams, FormFieldMetadata, handle_change_function_type } from '../../components/form-field';
-import { Form } from '../../components/form';
+import { FormFieldParams, FormFieldMetadata, handle_change_function_type } from '../../components/forms/form-field';
+import { Form } from '../../components/forms/form';
 import { LinkControl, function_VA } from '../../components/link-control';
 import { accounts_validate_null_input, accounts_validate_email, accounts_validate_password } from '../../services/validation-service';
 import { RequestStateInterface, NotificationStateInterface, GenericNullKeyArray } from '../../models/types';
