+++ website/src/pages/accounts/login-account.tsx
@@ -2,9 +2,9 @@ import * as React from 'react';
 import { Notification, NOTIFICATION_STYLE_ERROR } from '../../components/notification';
 import AccountsPageContainer from '../../components/accounts-page-container';
 import { Typography, Button, Grid } from '@material-ui/core';
-import { Form } from '../../components/form';
+import { Form } from '../../components/forms/form';
 import { LoginAccountPOST } from '../../models/http-requests';
-import { FormFieldParams, handle_change_function_type, FormFieldMetadata } from '../../components/form-field';
+import { FormFieldParams, handle_change_function_type, FormFieldMetadata } from '../../components/forms/form-field';
 import { GetHttpRequestDisplayName, http_post, LOGIN_ACCOUNT_END_POINT, HTTP_SUCCESS } from '../../services/http-service';
 import { LinkControl } from '../../components/link-control';
 import { RequestStateInterface, NotificationStateInterface, GenericNullKeyArray } from '../../models/types';
