--- website/src/pages/create-account/create-account.css
@@ -1,12 +1,0 @@
-.container {
-    min-height: 100vh;
-    padding: 10vw;
-}
-
-.group-container {
-    min-width: 20vw;
-}
-
-.control-container {
-    padding-top: 2vh;
-}
